# ARROW-17135: [C++] Reduce code size in compute/kernels/scalar_compare.cc