--- a/cpp/tools/binary_symbol_explore.py
+++ b/cpp/tools/binary_symbol_explore.py
@@ -114,6 +114,7 @@ if __name__ == '__main__':
                                                   'contender', 'diff'])
         pd.options.display.max_rows = 1000
         pd.options.display.max_colwidth = 150
+        print(diff[diff['diff'] < - 700])
         print(diff[diff['diff'] > 700])
     else:
         # TODO