@wesm 2022-07-20
Some more insight on symbols that got bigger/smaller in this change using tools/binary_symbol_explore.py:

https://gist.github.com/wesm/98eb3f28ce2a6ea134fa0f6d96208ce4

I don't really get why the applicator templates were generating ~40KB of code for single instantiations (like uint64 <= uint64), but probably our approach (so much for "zero cost abstractions") for generalizing to abstract between writing to an array versus packing a bitmap is causing too much code to be generated.