@pitrou 2022-07-20
Object file size:
* before:
```
$ size build-release/src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/scalar_compare.cc.o
   text	   data	    bss	    dec	    hex	filename
 458472	   6200	   1120	 465792	  71b80	build-release/src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/scalar_compare.cc.o
```
* after:
```
$ size build-release/src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/scalar_compare.cc.o
   text	   data	    bss	    dec	    hex	filename
 412088	   6336	   1120	 419544	  666d8	build-release/src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/scalar_compare.cc.o
```