@pitrou 2022-07-20
> [](https://gist.github.com/wesm/98eb3f28ce2a6ea134fa0f6d96208ce4)
> 
> I don't really get why the applicator templates were generating ~40KB of code for single instantiations (like uint64 <= uint64), but probably our approach (so much for "zero cost abstractions") for generalizing to abstract between writing to an array versus packing a bitmap is causing too much code to be generated.

That sounds quite outlandish :-o. Perhaps you can try disassembling one of those symbols to see if there's anything particularly fishy?