@wesm 2022-07-20
> That sounds quite outlandish :-o. Perhaps you can try disassembling one of those symbols to see if there's anything particularly fishy?

I tried to disassemble this and find out what's wrong on -- it appears to be a case of "-O3 behaving badly" but not sure why exactly -- the -O2 version of the function was < 2K where the -O3 version was ~40K. 

https://gist.github.com/wesm/ba184ec20ae85962d319a8ad2c3c421a