@pitrou 2022-07-20
This looks like auto-vectorization and loop unrolling gone completely crazy?
https://gist.github.com/wesm/ba184ec20ae85962d319a8ad2c3c421a#file-bad-dis-L2105

I wonder what would happen if we compiled the entire codebase with `-O2` instead of `-O3`...