@wesm 2022-07-20
Yeah I am thinking that we should stop using -O3 altogether