@wesm 2022-07-20
+1. I'll start an e-mail thread about the optimization level setting