@wesm 2022-07-26
That's a really weird valgrind error. I'll try to reproduce it locally and see what I can find